Melanoma_Detection_Tool/backend/requirements.txt
fastapi
uvicorn
streamlit
tensorflow
pillow
requests
python-multipart